requirements.txt
certifi==2018.4.16
chardet==3.0.4
idna==2.7
pyserial==3.1.1
requests==2.19.1
urllib3==1.23
websocket-client
